Update BudgetImageMatching.py
Put in a rough draft for the resize using cv2.resize(), need to look into if we want to test multiple resize methods to see if we can't find a match on the image by sending it through a different scaling method just to see what happens. Could also simply try to justify the differences if they fall within reasonable error margins.