Experimental rand-edge detection
Added a new experimental mode to the Canny Edge detector. This mode is experimental in that I sample the random points and arrange them into a new image, which is 1 x N, where N is the number of points sampled. The Canny edge detector is then run over it, resulting in our images to compare. It reports about 3.3 to 3.5% difference between the images in this 'format', but it also has a large number of issues when comparing two different images. Especially since it looks for edges between the randomly sampled points, so it will generally result in too many false positives. 

This is just the first of some experimental mode ideas being bounced around, and it's performance is pretty abysmal, so it won't see much use, despite the time-save.